renamed git shared folder
