+++ internal/shared/git/git_test.go
@@ -1,4 +1,4 @@
-package gitconfig
+package git
 
 import (
 	"context"
